Alright then, i32

+++ 2022/Rust/src/d18.rs
@@ -2,18 +2,18 @@ use std::cmp::max;
 use std::collections::HashSet;
 use crate::tools;
 
-pub fn xyz_to_index(x : u32, y : u32, z : u32, dims : (u32, u32, u32)) -> u32 {
+pub fn xyz_to_index(x : i32, y : i32, z : i32, dims : (i32, i32, i32)) -> i32 {
   x + (y * dims.0) + (z * dims.0 * dims.1)
 }
 
-pub fn bv(z : bool) -> u32 {
+pub fn bv(z : bool) -> i32 {
     if z { 1 } else { 0 }
 }
 
 pub struct Gubbins {
-    cubes : Vec<(u32, u32, u32)>,
-    map : HashSet<u32>,
-    dims : (u32, u32, u32)
+    cubes : Vec<(i32, i32, i32)>,
+    map : HashSet<i32>,
+    dims : (i32, i32, i32)
 }
 
 pub fn parse(input : String) -> Gubbins {
@@ -25,7 +25,7 @@ pub fn parse(input : String) -> Gubbins {
         }
         let mut xyz = Vec::new();
         for coord in cube.split(',') {
-            xyz.push(coord.parse::<u32>().unwrap());
+            xyz.push(coord.parse::<i32>().unwrap());
         }
         cubes.push((xyz[0] + 1, xyz[1] + 1, xyz[2] + 1));
         let xyz= cubes.last().unwrap();
@@ -41,23 +41,23 @@ pub fn parse(input : String) -> Gubbins {
     Gubbins{ cubes, map, dims}
 }
 
-pub fn part1(cubes : &[(u32, u32, u32)], map : &HashSet<u32>, dims : &(u32, u32, u32)) -> u32 {
+pub fn part1(cubes : &[(i32, i32, i32)], map : &HashSet<i32>, dims : &(i32, i32, i32)) -> i32 {
     let deltas: [[i32 ; 6] ; 3] = [[-1, 1, 0, 0, 0, 0], [0, 0, -1, 1, 0, 0], [0, 0, 0, 0, -1, 1]];
     let mut total = 0;
     for cube in cubes.iter() {
         for face in 0..6 {
-            let nx = cube.0 as i32 + deltas[0][face];
-            let ny = cube.1 as i32 + deltas[1][face];
-            let nz = cube.2 as i32 + deltas[2][face];
+            let nx = cube.0 + deltas[0][face];
+            let ny = cube.1 + deltas[1][face];
+            let nz = cube.2 + deltas[2][face];
 
             total += bv(!map.contains(
-                &xyz_to_index(nx as u32, ny as u32, nz as u32, *dims)));
+                &xyz_to_index(nx, ny, nz, *dims)));
         }
     }
     total
 }
 
-pub fn part2(map : &HashSet<u32>, dims : &(u32, u32, u32)) -> u32 {
+pub fn part2(map : &HashSet<i32>, dims : &(i32, i32, i32)) -> i32 {
     let deltas: [[i32 ; 6] ; 3] = [[-1, 1, 0, 0, 0, 0], [0, 0, -1, 1, 0, 0], [0, 0, 0, 0, -1, 1]];
     let mut queue = Vec::new();
     let mut visited = HashSet::new();
@@ -67,17 +67,17 @@ pub fn part2(map : &HashSet<u32>, dims : &(u32, u32, u32)) -> u32 {
     while !queue.is_empty() {
         let cube = queue.pop().unwrap();
         for face in 0..6 {
-            let nx = cube.0 as i32 + deltas[0][face];
-            let ny = cube.1 as i32 + deltas[1][face];
-            let nz = cube.2 as i32 + deltas[2][face];
+            let nx = cube.0 + deltas[0][face];
+            let ny = cube.1 + deltas[1][face];
+            let nz = cube.2 + deltas[2][face];
             if nx < 0 || ny < 0 || nz < 0 { continue; };
-            if nx >= dims.0 as i32 || ny >= dims.1 as i32 || nz >= dims.2 as i32 { continue; }
+            if nx >= dims.0 || ny >= dims.1 || nz >= dims.2 { continue; }
 
-            let index = xyz_to_index(nx as u32, ny as u32, nz as u32, *dims);
+            let index = xyz_to_index(nx, ny, nz, *dims);
             if map.contains( &index) {
                 total += 1;
             } else if !visited.contains(&index) {
-                queue.push((nx as u32, ny as u32, nz as u32));
+                queue.push((nx, ny, nz));
                 visited.insert(index);
             }
         }
@@ -86,12 +86,12 @@ pub fn part2(map : &HashSet<u32>, dims : &(u32, u32, u32)) -> u32 {
     total
 }
 
-pub fn _solve(input : String) -> (u32, u32) {
+pub fn _solve(input : String) -> (i32, i32) {
     let d = parse(input);
     (part1(&d.cubes, &d.map, &d.dims),part2(&d.map, &d.dims))
 }
 
-pub fn solve() -> (u32, u32) {
+pub fn solve() -> (i32, i32) {
     _solve(tools::read_file_contents(&tools::find_input_path("18")))
 }
